beauty

+++ SOnA/Domain/Model/MovieScreening.cs
@@ -3,10 +3,10 @@
     public class MovieScreening
     {
         public DateTime dateAndTime { get; private set; }
-        public double pricePerSeat { get; private set; }
+        public decimal pricePerSeat { get; private set; }
         public Movie movie { get; private set; }
 
-        public MovieScreening(DateTime dateAndTime, double pricePerSeat, Movie movie)
+        public MovieScreening(DateTime dateAndTime, decimal pricePerSeat, Movie movie)
         {
             this.dateAndTime = dateAndTime;
             this.pricePerSeat = pricePerSeat;
