+++ SOnA/Domain/Model/MovieTicket.cs
@@ -14,7 +14,7 @@ public class MovieTicket
         this.movieScreening = movieScreening;
     }
 
-    public double Price()
+    public decimal Price()
     {
         return this.movieScreening.pricePerSeat;
     }
